requirements.txt
coverage==4.4.2
dj-database-url==0.4.2
Django==1.11.9
django-crispy-forms==1.7.0
django-debug-toolbar==1.11
django-widget-tweaks==1.4.1
Markdown==3.1.1
Pillow==5.0.0
python-decouple==3.1
pytz==2017.3
sqlparse==0.3.0
